src/test/java/com/refactor/captcha/NodeTest.java
package com.refactor.captcha;

import org.junit.Test;

import static org.junit.Assert.assertEquals;


public class NodeTest {

    @Test
    public void display_ValueNode_printsTheValue() {
        Node valueNode = new Node(1);
        assertEquals("1", valueNode.display());
    }

    @Test
    public void display_OperatorNode_printsTheEquation() {
        Node operatorNode = new Node("+", new Node(1), new Node(2));
        assertEquals("1 + 2", operatorNode.display());
    }

    @Test
    public void display_complexOperatorNodes() {
        Node f = new Node("*", new Node(2), new Node(5));
        Node g = new Node("+", f, new Node(3));

        assertEquals("2 * 5 + 3", g.display());
    }

    @Test
    public void compute_ValueNode_givesTheValueBack() {
        Node valueNode = new Node(2);
        assertEquals(2, valueNode.compute().intValue());
    }

    @Test
    public void compute_OperatorNode_givesTheComputedResult() {
        Node opsNode = new Node("-", new Node(5), new Node(3));
        assertEquals(2, opsNode.compute().intValue());
    }

    @Test
    public void compute_complexOperatorNodes() {
        Node f = new Node("*", new Node(2), new Node(5));
        Node g = new Node("+", f, new Node(3));

        assertEquals(13, g.compute().intValue());
    }
}
